Rename the underscore-oop module to underscore-array-methods
